Prefer @type to @opaque (and fix dialyzer)
See here: https://elixirforum.com/t/difference-between-type-and-opaque/41611/4

Because matchers will want to receive
the wrapped expected value, and the matchers are necessarily defined outside of
the Expect module, the typespec for Expect needs to be non-opaque, as @opaque
enforces it to only ever be used by the module itself, although the dialyzer
warning for this is (as always) obscure and misleading.

I suppose we could bring this back by
changing the interface for matchers, to not pass the expect itself, and to instead
pass the value itself and optionally some args, but I'd like to wait to do that until
I know better what the interface wants to be with the matchers.

Plus there is already a breaking change for matchers that I intend to make soon.